Adding an intensive model
Added a goodwin
